feat: remove audited methods from api client and add them directly to the pipeline
