+++ eox_nelp/api_clients/pearson_rti.py
@@ -11,14 +11,6 @@ from django.conf import settings
 from eox_nelp.api_clients import AbstractSOAPClient
 from eox_nelp.api_clients.authenticators import PKCS12Authenticator
 
-try:
-    from eox_audit_model.decorators import audit_method
-except ImportError:
-    def audit_method(action):  # pylint: disable=unused-argument
-        """Identity audit_method"""
-        return lambda x: x
-
-
 LOGGER = logging.getLogger(__name__)
 
 
@@ -72,45 +64,40 @@ class PearsonRTIApiClient(AbstractSOAPClient):
         Returns:
             <Dict>: Parsed xml response to Dict format.
         """
-        @audit_method(action="Import Candidate Demographics")
-        def import_candidate_demographics_request(payload):
-            """This is a wrapper that allows to make audit-able the import_candidate_demographics method."""
-            xml_response = BeautifulSoup(
-                self.make_post("cxfws2/services/CDDService", payload),
-                "xml",
-            )
-            fault = xml_response.find("soapenv:Fault")
+        xml_response = BeautifulSoup(
+            self.make_post("cxfws2/services/CDDService", payload),
+            "xml",
+        )
+        fault = xml_response.find("soapenv:Fault")
 
-            # There are multiple kind of errors, some of them generates a soapenv:Fault section and others
-            # generates an status error section, the following condition handles the first case.
-            if fault:
-                return {
-                    "status": "error",
-                    "fault_code": xml_response.find("faultcode").text,
-                    "message": xml_response.find("faultstring").text,
-                }
-
-            status = xml_response.find("status")
-
-            if status:
-                return {
-                    "status": status.text.lower(),
-                    "message": xml_response.find("message").text,
-                    "candidate_id": xml_response.find("cdd:cddResponse").attrs.get("candidateID"),
-                    "client_candidate_id": xml_response.find("cdd:cddResponse").attrs.get("clientCandidateID"),
-                }
-
-            LOGGER.error(
-                "An unexpected error has occurred trying to make a CDD request getting the following response: %s",
-                xml_response,
-            )
-
+        # There are multiple kind of errors, some of them generates a soapenv:Fault section and others
+        # generates an status error section, the following condition handles the first case.
+        if fault:
             return {
-                "status": "unexpected error",
-                "response": xml_response,
+                "status": "error",
+                "fault_code": xml_response.find("faultcode").text,
+                "message": xml_response.find("faultstring").text,
             }
 
-        return import_candidate_demographics_request(payload)
+        status = xml_response.find("status")
+
+        if status:
+            return {
+                "status": status.text.lower(),
+                "message": xml_response.find("message").text,
+                "candidate_id": xml_response.find("cdd:cddResponse").attrs.get("candidateID"),
+                "client_candidate_id": xml_response.find("cdd:cddResponse").attrs.get("clientCandidateID"),
+            }
+
+        LOGGER.error(
+            "An unexpected error has occurred trying to make a CDD request getting the following response: %s",
+            xml_response,
+        )
+
+        return {
+            "status": "unexpected error",
+            "response": xml_response,
+        }
 
     def import_exam_authorization(self, payload):
         """The EAD service imports exam authorization data into the Pearson VUE system. This
@@ -122,41 +109,36 @@ class PearsonRTIApiClient(AbstractSOAPClient):
         Returns:
             <Dict>: Parsed xml response to Dict format.
         """
-        @audit_method(action="Import Exam Authorization")
-        def import_exam_authorization_request(payload):
-            """This is a wrapper that allows to make audit-able the import_exam_authorization method."""
-            xml_response = BeautifulSoup(
-                self.make_post("cxfws2/services/EADService", payload),
-                "xml",
-            )
-            fault = xml_response.find("soapenv:Fault")
+        xml_response = BeautifulSoup(
+            self.make_post("cxfws2/services/EADService", payload),
+            "xml",
+        )
+        fault = xml_response.find("soapenv:Fault")
 
-            # There are multiple kind of errors, some of them generates a soapenv:Fault section and others
-            # generates an status error section, the following condition handles the first case.
-            if fault:
-                return {
-                    "status": "error",
-                    "fault_code": xml_response.find("faultcode").text,
-                    "message": xml_response.find("faultstring").text,
-                }
-
-            status = xml_response.find("status")
-
-            if status:
-                return {
-                    "status": status.text.lower(),
-                    "message": xml_response.find("message").text if xml_response.find("message") else "",
-                    "client_candidate_id": xml_response.find("clientCandidateID").text,
-                }
-
-            LOGGER.error(
-                "An unexpected error has occurred trying to make a EAD request getting the following response: %s",
-                xml_response,
-            )
-
+        # There are multiple kind of errors, some of them generates a soapenv:Fault section and others
+        # generates an status error section, the following condition handles the first case.
+        if fault:
             return {
-                "status": "unexpected error",
-                "response": xml_response,
+                "status": "error",
+                "fault_code": xml_response.find("faultcode").text,
+                "message": xml_response.find("faultstring").text,
             }
 
-        return import_exam_authorization_request(payload)
+        status = xml_response.find("status")
+
+        if status:
+            return {
+                "status": status.text.lower(),
+                "message": xml_response.find("message").text if xml_response.find("message") else "",
+                "client_candidate_id": xml_response.find("clientCandidateID").text,
+            }
+
+        LOGGER.error(
+            "An unexpected error has occurred trying to make a EAD request getting the following response: %s",
+            xml_response,
+        )
+
+        return {
+            "status": "unexpected error",
+            "response": xml_response,
+        }
