+++ eox_nelp/pearson_vue/pipeline.py
@@ -25,6 +25,13 @@ from eox_nelp.pearson_vue.constants import PAYLOAD_CDD, PAYLOAD_EAD, PAYLOAD_PIN
 from eox_nelp.pearson_vue.utils import generate_client_authorization_id, update_xml_with_dict
 from eox_nelp.signals.utils import get_completed_and_graded
 
+try:
+    from eox_audit_model.decorators import audit_method
+except ImportError:
+    def audit_method(action):  # pylint: disable=unused-argument
+        """Identity audit_method"""
+        return lambda x: x
+
 logger = logging.getLogger(__name__)
 
 User = get_user_model()
@@ -179,26 +186,32 @@ def import_candidate_demographics(cdd_request, **kwargs):  # pylint: disable=unu
     Raises:
         Exception: If the Pearson VUE RTI service does not accept the import request.
     """
-    api_client = PearsonRTIApiClient()
-    payload = {
-        "soapenv:Envelope": {
-            "soapenv:Header": {
-                "wsse:Security": {
-                    "wsse:UsernameToken": {
-                        "wsse:Username": getattr(settings, "PEARSON_RTI_WSDL_USERNAME"),
-                        "wsse:Password": {
-                            "#text": getattr(settings, "PEARSON_RTI_WSDL_PASSWORD")
+    @audit_method(action="Import Candidate Demographics")
+    def import_candidate_demographics_request(cdd_request):
+        """This is a wrapper that allows to make audit-able the import_candidate_demographics method."""
+        api_client = PearsonRTIApiClient()
+        payload = {
+            "soapenv:Envelope": {
+                "soapenv:Header": {
+                    "wsse:Security": {
+                        "wsse:UsernameToken": {
+                            "wsse:Username": getattr(settings, "PEARSON_RTI_WSDL_USERNAME"),
+                            "wsse:Password": {
+                                "#text": getattr(settings, "PEARSON_RTI_WSDL_PASSWORD")
+                            },
                         },
                     },
                 },
+                "soapenv:Body": {
+                    "sch:cddRequest": cdd_request
+                },
             },
-            "soapenv:Body": {
-                "sch:cddRequest": cdd_request
-            },
-        },
-    }
-    payload = update_xml_with_dict(PAYLOAD_CDD, payload)
-    response = api_client.import_candidate_demographics(payload)
+        }
+        payload = update_xml_with_dict(PAYLOAD_CDD, payload)
+
+        return api_client.import_candidate_demographics(payload)
+
+    response = import_candidate_demographics_request(cdd_request)
 
     if response.get("status", "error") != "accepted":
         # pylint: disable=broad-exception-raised
@@ -224,26 +237,31 @@ def import_exam_authorization(ead_request, **kwargs):  # pylint: disable=unused-
     Raises:
         Exception: If the Pearson VUE RTI service does not accept the import request.
     """
-    api_client = PearsonRTIApiClient()
-    payload = {
-        "soapenv:Envelope": {
-            "soapenv:Header": {
-                "wsse:Security": {
-                    "wsse:UsernameToken": {
-                        "wsse:Username": getattr(settings, "PEARSON_RTI_WSDL_USERNAME"),
-                        "wsse:Password": {
-                            "#text": getattr(settings, "PEARSON_RTI_WSDL_PASSWORD")
+    @audit_method(action="Import Exam Authorization")
+    def import_exam_authorization_request(ead_request):
+        """This is a wrapper that allows to make audit-able the import_exam_authorization method."""
+        api_client = PearsonRTIApiClient()
+        payload = {
+            "soapenv:Envelope": {
+                "soapenv:Header": {
+                    "wsse:Security": {
+                        "wsse:UsernameToken": {
+                            "wsse:Username": getattr(settings, "PEARSON_RTI_WSDL_USERNAME"),
+                            "wsse:Password": {
+                                "#text": getattr(settings, "PEARSON_RTI_WSDL_PASSWORD")
+                            },
                         },
                     },
                 },
+                "soapenv:Body": {
+                    "sch:eadRequest": ead_request
+                },
             },
-            "soapenv:Body": {
-                "sch:eadRequest": ead_request
-            },
-        },
-    }
-    payload = update_xml_with_dict(PAYLOAD_EAD, payload)
-    response = api_client.import_exam_authorization(payload)
+        }
+        payload = update_xml_with_dict(PAYLOAD_EAD, payload)
+        return api_client.import_exam_authorization(payload)
+
+    response = import_exam_authorization_request(ead_request)
 
     if response.get("status", "error") != "accepted":
         # pylint: disable=broad-exception-raised
